requirements/dev.txt
-r base.txt

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
pylint==3.0.2
mypy==1.7.1

# Debug
ipython==8.17.2
ipdb==0.13.13
